table: fix the Originator attribute update for route-reflector client
Needs to set the Originator attribute to be to the router-id of the
advertiser, not ours.

RFC 4684 3.2 says:

When advertising RT membership NLRI to a route-reflector client, the
Originator attribute shall be set to the router-id of the advertiser,
and the Next-hop attribute shall be set of the local address for that
session.

Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>